linux: use rtnetlink directly
This change uses rtnetlink directly per route query instead of relying
on the routing information base.

This fixes a routing issue where, with some VPNs (wireguard based ones),
this library would return the incorrect source IP when queried.

This is because Wireguard creates a new table and routing rules that are
not reflected in the routing information base.

Instead of trying to recreate this logic, we can query kernel directly
via the rtnetlink socket. This is a bit painful from Go, but doable.

Consult `man 7 rtnetlink` for more information on the rtnetlink
interface.